# swagger_server/models/data_delivery_mode.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class DataDeliveryMode(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    """
    allowed enum values
    """
    ONE_TIME = "ONE_TIME"
    CONTINUOUS = "CONTINUOUS"
    def __init__(self):  # noqa: E501
        """DataDeliveryMode - a model defined in Swagger

        """
        self.swagger_types = {
        }

        self.attribute_map = {
        }

    @classmethod
    def from_dict(cls, dikt) -> 'DataDeliveryMode':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The DataDeliveryMode of this DataDeliveryMode.  # noqa: E501
        :rtype: DataDeliveryMode
        """
        return util.deserialize_model(dikt, cls)